Debug script to find out why analysis stops on stallo
When an uncaught exception is raised in analyse_corpus, the script will
hang if ran in parallel mode (the default).

This script will run both the conversion and analysis process without using
the parallelism offered by the multiprocessing module and will stop when an
uncaught exception is raised, making it possible to see what stops either the
conversion or analysis.