Three new tests
1. Test that a new basename is computed if one tries to move a parallel
file to an existing file that is not a duplicate of the parallel file
2. Test that a parallel file that one tries to move to an existing file
name really is moved to a new basename and that the metadata also is
updated accordingly
3. Test that an exception is raised when trying to move a parallel file
to an existing file, when that existing file is a duplicate of the
parallel file. Also check that no other files in the change set is
changed or moved when this happens.