Pythonify korp_scripts/update_mono
This commit adds korp_scripts/update_mono/compile_cwb_files.py,
which is a re-write of the functionality of all the scripts in that folder.
It enables the entire process to happen automatically, and uses only the
python standard library (so no dependencies). There's no dependency on
java anymore either (to run the compile_corpus.xsl file).

Still a WIP (isn't everything..), so there are probably a few things that
the old scripts does that I have yet to add, but in principle it should
work already. Tested succesfully locally with a smaller corpus.

After running the script, .yaml files need to be added to the Korp
backend configuration directory CORPUS_CONFIG/corpora.
One .yaml file for each file in the created registry/ folder.
It contains information to Korp for how to present the corpus in the
web interface. Things like "description" and such go in there. Refer to the
Korp documentation for more information about that.
This only applies to v9 of Korp, as previously configuration was done
differently.

Some of these things could be done programmatically, but ultimately there
are settings in there that we cannot determine from this script (at least
not trivally), so automating it further is probably not worth it. As long
as the documentation is good, it's really ok.